Test status: passed all

+++ node.c
@@ -6,7 +6,7 @@
 /*   By: rosousa- <rosousa-@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/12/21 02:33:26 by rosousa-          #+#    #+#             */
-/*   Updated: 2026/01/18 15:58:56 by rosousa-         ###   ########.fr       */
+/*   Updated: 2026/01/18 17:55:27 by rosousa-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -41,7 +41,7 @@ void	*add_node_back(t_list **head_a, t_list *node)
 	return (*head_a);
 }
 
-void	put_content(int len_array, int *ptr, t_list **head_a)
+void	put_content(int len_array, long *ptr, t_list **head_a)
 {
 	t_list *node;
 	int i;
@@ -51,7 +51,7 @@ void	put_content(int len_array, int *ptr, t_list **head_a)
 		return;
 	while (i < len_array)
 	{
-		node = new_node(ptr[i]);
+		node = new_node((int)ptr[i]);
 		if (!node)
 		{
 			if (i == 0)
